refactor(watch): move down resources list to a suspended RSC
